Extraxt metric agent service section
